requirements.txt
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
web3==6.11.3
py-solc-x==2.0.2
python-docx==1.1.0
Pillow==10.1.0
scipy==1.11.4

